add NPC api
