+++ js/render/core/npc.md
@@ -1,0 +1,70 @@
+# Documentation for npc.js
+
+A sample usage can be found in scenes/spider.js.
+
+## Class `NPCSystem`
+
+### `new NPCSystem(model, dimX, dimZ)`
+
+Optional:  
+`dimX` and `dimZ` define the dimension of the terrain system. Default is 16x16.  
+
+This will create a NPC System, which is the prerequisite for the NPCs and terrain system.
+
+### `.initRender(node, meshTerrainName)`
+
+Optional:  
+`meshTerrainName`. To render the terrain, a combined mesh will be generated, this name is for the mesh. Default is "_NPCTerrain".
+
+Please call this function before calling any other function related to the rendering.
+
+### `.getRootNode()`
+
+returns the root node of the NPC System, which can be used to move, scale, etc. the whole system.
+
+### `.getTerrain()`
+
+returns the NPCTerrain object.
+
+### `.addNPC(type, id, render)`
+
+adds a new NPC to the system.  
+Available types are: "triped", "quadruped", "pentaped", "hexapod", "heptapod", "octopod".
+
+Optional:  
+`id`. The id of the NPC. If not specified, an id will be autogenerated.  
+`render`. A boolean value to control if npc.initRender() is called. When `false`, the NPC is not added to the rendering. Default is `true`.
+
+### `.getNPC(id)`
+
+returns the NPC object by id.
+
+### `.setNPCMoveVec(id, moveVec)`
+
+sets the vector of the movement of the NPC with certain id. It is used to control the NPC.
+
+### `.update()`
+
+Call this in the model.animate(()=>{}) loop to update the NPC as well as their rendering if applicable.
+
+## Class `NPCTerrain`
+
+### `.clear()`
+
+resets the terrain to a plain at y=0 (local).
+
+### `.randomize()`
+
+generate the terrain with noise-based height. Data only: it will NOT update the mesh. To update the mesh, call `.generateMesh()` after it.
+
+### `.generateMesh(meshName)`
+
+updates the mesh of terrain.  
+Optional:  
+`meshName`. Default is "_NPCTerrain".
+
+## Class `RobotMultiLegs`
+
+### `.resetPosY()`
+
+resets the robot's position in the Y axis to fit the terrain. Highly recommended after a terrain re-generation.
